Moved custom icons related calls
